jsonrl, ion: fix conditional in ion.Chunker.compressOrFlush
Fix a conditional in ion.Chunker.compressOrFlush that incorrectly
led to us skipping a flush when the input was incompressible *and*
the current symbol table wasn't fully up-to-date *and* one of the
recent symbols was so large that it would cause the serialized data
to split a block boundary.

Also, update jsonrl.Convert to *not* call ion.Chunker.Flush after
each call. We call Flush in blockfmt.Converter.Run, and there isn't
any need to flush more frequently than that.