snellerd: fix unit test with -cpu=1
TestSimpleFS fails 100% of the time with -cpu=1.

Something about ingesting ion data via UnsafeION()
in the same block as ingesting json data produces
invalid output ion. This isn't a code path we actually
use in production (we don't support *.ion),
so the shortcut here is just to avoid tickling the bug in unit tests.