blockfmt: drop vm dependency
Importing sneller/vm means you incur a bit of start-up
overhead allocating the vm memory area (not to mention
the binary size bloat). The blockfmt package only
depended on vm for allocating pages to send to the vm,
but we can stick that behind a layer of indirection
so that using blockfmt from sneller/db doesn't automatically
imply importing the vm package as a hidden dependency.