db: add etag check when flushing index
Track the 'latest' index etag in db.IndexCache
and confirm that it hasn't changed before flushing
index files.

This isn't 100% airtight because there is TOCTOU race
between checking the old file and overwriting it, but
it can at least help us detect when our mutual exclusion
mechanisms have failed to provide the guarantees we want.

(NOTE: this protection doesn't currently apply to
the scan path because the scan path doesn't use an index cache...)