Fixed a possible issue in bcunboxts regarding unboxing microseconds
This fix makes sure that we would only gather active lanes that
actually have microseconds in a boxed timestamp data. Gathering
microseconds requires to read additional 4 bytes of a boxed data.
If we read blindly, we could be reading non-active lanes or timestamp
data that don't have microseconds. The biggest problem is non-active
lanes as the slices representing the boxed value could be completely
dirty and could contain anything.