blockfmt: filetree: rework eviction strategy
Currently, we evict all the filetree leaf nodes on each tree sync.
For very large trees, this is suboptimal: "hot" insert paths end
up requiring that we fault in the same leaf nodes each time we
insert a batch.

Have Filetree.sync manage eviction instead: as it walks the tree,
it evicts non-dirty nodes, and rewrites dirty nodes (after storing them)
to store only their compressed representation. This means that "hot"
nodes only have to do a bit of decompression to perform an insert,
rather than having to go all the way to the backing store.

One drawback here is that the amount of memory the tree can occupy
is only bound by the number of inserts we perform, so we should keep
tabs on memory use (and peek # inserts) to ensure this strategy isn't
too aggressive.