expr: introduce StringMatch
We have a bunch of operators that require a constant
string on the right-hand-side and optionally accept
an ESCAPE parameter. Let's call these "StringMatch"
rather than "Comparison."

No major semantic changes other than that the ESCAPE
string length and the regexp validity are both checked
in expr.Check(), so we should catch errors there earlier.