vm: symtab: snapshot and restore on Unmarshal
Right now vm.rowSplitter uses two symbol tables:
the first is used to keep track of the state of
the *input* symbol table as it is appended to,
and the second tracks the state of the *output*
symbol table of the query (which, at any point in time,
is a super-set of the input symbol table).
We needed to track these separately so that new symbols
arriving in the input stream would cause the output symbol
table to be invalidated and for symbolize() to be called
to re-populate the output table.

We can achieve the same effect with just one symbol table
by keeping a "snapshot" of the state of the input symbol
table and then rewinding to that state before processing
new input symbols. This is pretty easy in theory, but because
vm.symtab manages a bunch of vm-allocated memory, the memory
allocator also needs to support snapshot+restore semantics.
That's the purpose of vm.slab.

Since the new zion-accelerated COUNT(*) operation touches
so few bytes, it spent a lot of time dealing with symbol tables.
Moving to one symbol table makes it 20% faster:
```
~/sneller-core/oss/cmd/sneller $ GOMAXPROCS=1 ./sneller -S -j "select count(*) from '~/1day.zion'"
{"count": 3259519}
120GiB/s (scanned 9780068352 B (9.11 GB) in 79.427723ms), allocated memory: 1700456 B (1.62 MB), allocations: 28138
~/sneller-core/oss/cmd/sneller $ GOMAXPROCS=1 ./sneller -S -j "select count(*) from '~/1day.zion'"
{"count": 3259519}
120GiB/s (scanned 9780068352 B (9.11 GB) in 79.570368ms), allocated memory: 1715208 B (1.64 MB), allocations: 28141
~/sneller-core/oss/cmd/sneller $ GOMAXPROCS=1 ./sneller-old -S -j "select count(*) from '~/1day.zion'"
{"count": 3259519}
98.3GiB/s (scanned 9780068352 B (9.11 GB) in 97.135721ms), allocated memory: 1750000 B (1.67 MB), allocations: 28172
~/sneller-core/oss/cmd/sneller $ GOMAXPROCS=1 ./sneller-old -S -j "select count(*) from '~/1day.zion'"
{"count": 3259519}
98.9GiB/s (scanned 9780068352 B (9.11 GB) in 96.605036ms), allocated memory: 1774736 B (1.69 MB), allocations: 28223
```