partiql: fix lexing numbers
If we find the number beginning (a digit, the '.' or '-'),
then treat all characters from that char until a whitespace
or operator as a possible number representation. Do not
perform much work during scanning, let the strconv
procedures to detected syntax and other errors.

Also extend lexing to support the sign characters in the
engineering notation; now both '1e+5' and '8e-9' are valid
(and are parsed correctly by strconv.ParseFloat).

Additionally, fix lexing numbers glued with '-' operator.
A query like `SELECT 1-2` was rejected, while `SELECT 1 - 2`
was accepted. Tweak the lexer to emit proper tokens in such
cases.