vm: disable andprop optimization
The andprop optimization has been the source of a few subtle bugs
and its implementation is quite complex. For now, simply remove it.
We may be able to add back a simplified version in the future.

Also, remove TestCompileSSA. The TestQueries harness gets much
better coverage of SSA compilation at this point, and maintaining
these compilation tests is a pain (we don't particularly care that
the bytecode produced stays exactly the same; we only care that
it is correct!)

Fixes #2128