vm: fix inferred type of (I)LIKE expressions
The type of LIKE/ILIKE expressions in certain places
in ssa compilation was incorrectly inferred to be
string-typed, since that's the result type of the
actual SSA opcode. (*prog).serialized and (*prog).equals
were the two main offenders.

Compile LIKE/ILIKE to be wrapped in a trivial tuple
instruction that extracts only the mask part of the
opcode chain as the return value. Additionally, set
the value's notMissing status to the string-ness of
the original input argument to the chain implementing
the LIKE operation.

This fix reveals one issue in ILIKE that remains to be fixed.