vm: speed up benchmarking
The BenchmarkTestQueries harness takes a while to run
because it rebuilds the input dataset each time the
Go benchmark harness decides to re-run the benchmark
with a different b.N value.

Make sure the input dataset is only computed once,
and only when the sub-benchmark is actually selected
to run (i.e. via -bench ...)