blockfmt: compact objects in Index.SyncOutputs
Change the behavior of SyncOutputs to compact object lists
before sticking them into the IndirectTree.

One reason to delay compaction up to this point is that
the FileTree uses an integer packfile ID to map inputs
to packfiles, and doing compaction before sticking objects
into the IndirectTree would invalidate that mapping.
(It's not clear that we actually need to maintain this
mapping in practice, though.)

Once we have accumulated some number of objects in Index.Inline,
the oldest half of the inline list is compacted and then stuck
into the indirect tree.