expr: extend hints with a set of possible values
Currently a hint may be used to determine the type set of a node.
This commit adds an ability to return a finite set of possible
values the given node may have. This information can be used
to simplify comparisons with constants as well as the `IN`
expressions.

In practise we will obtain the set of values from a partitioning
schema, but this change does not assume anything about the
source of information.

Note: simplification does not utilize the fact that a set may
contain single element and then be inlined into the query. That
kind of simplification is meant to be a separate step of
optimization, likely very first one.