vm: improve findsym_tail performance
It is often the case that the first vpgatherdd
in the findsym loop can grab all of the field symbol
varint bits *and* all the value descriptor bits,
even when the value has a varint-encoded length.
Instead of issuing a second gather unconditionally
when the value has a uvarint-encoded length, first
check to see if the uvarint stop bit is already in
the source data register, and if so, skip loading more data.

Also, hoist some constant broadcasts out of the core loop.