cmd/sdb: refactor applets
Move applet definitions into separate files and
move some of the flags that are applet-specific
(i.e. -o, -m, etc.) to flag sets that are parsed
within the applet itself. Hopefully this will keep the
proliferation of sub-commands from cluttering the top-level
documentation when trying to discover some of the tool's functionality.

Also, move -build and -version into "buildinfo" and "version" applets
along with a "help" applet that works similarly to "-h".