plan/pir: implement inner equi-joins
Add very basic equi-join support to the query planner.

The way this works is that we introduce an EquiJoin pir.Step
that exists for a little while before optimization in order
for it to absorb some predicate and dereference push-down information.
Then, part way through optimization, we eliminate the EquiJoin step(s)
altogether by converting them into sub-queries plus HASH_REPLACEMENT()
operations.

As-is, this has some shortcomings:

 - We're using HASH_REPLACEMENT(), which right now will
   cause run-time errors if the build side of the join is
   more than 10,000 records.

 - We naively assign the left-most table as the probe table
   and the other table(s) as build tables. This may be far from ideal.

 - Like cross-joins, all the fields in the query need to be fully-qualified,
   and we do not (yet) support `SELECT * ...` in the context of a join.