blockfmt: reduce Converter parallelism
blockfmt.MultiWriter lets us convert multiple input files
in parallel, but if we naively use it every time there is
more than one input file, then we run the risk of creating
under-full blocks with lots of nop padding.

Instead, use the number of input bytes to pick a reasonable
level of parallelism. In practice, ingest performance tends
to be bound by the GET latency from the backing store, and both
the single-stream and multi-stream conversion paths have pretty
deep prefetching. There's no real-world performance benefit
to multi-stream conversion when the inputs are very small.

For now, use the heuristic that we shouldn't exceed one stream
of conversion for each 2MB of data being converted at at time.
(For now, we assume anything that is compressed is 5x larger than
the stated object size, so we'd be using 1 CPU per ~400kB of gzipped data.)
In practice this means most inserts will be single-stream,
and we'll use the multi-stream path mostly for backfilling
or for tables with extremely high rates of ingest.