blockfmt: Converter: increase prefetch depth
Based on some benchmarking against S3, we need a way deeper
prefetch queue in order to get decent performance when reading
lots of small files.

For now, set the default parallelism to 400 (up from 64)
and expose it as an adjustable parameter in blockfmt.Converter
in case it needs to be tuned down the road.