aws/s3: perform async server-side copies
When we are copying using S3 server-side copy to
move data around, hide the latency of the copy-part
operation by performing it in the background.

The motivation for this change is that traces indicate
a lot of the sync latency for blockfmt.Converter.runSingle()
is in blockfmt.Converter.fastPrepend(), which performs a
server-side copy of a prefix to a packfile that we want
to avoid entirely re-ingesting. If we perform this copy in
the background, then we only have to eat the server-side upload latency
if it takes longer for that to complete than it does for us to actually
read in the new data and format it.