plan/pir: fix JOIN planning with CTEs
There was an old TODO sticking around in the build.go code
that described an issue I ran into when trying to run a JOIN
using CTEs: we would infer the wrong variable bindings when
a CTE was re-bound using an explicit table variable.

In order to allow explicit re-naming of CTEs that are flattened
into the parent trace, introduce a temporary pir.Step during construction
and then eliminate it once the trace has been built completely.

Also, fix an issue with tableReplacer that was preventing it from
correctly flattening CTE bindings in JOINs.