vm: store hash tables in slab memory
Since we'd like to be able to support much larger
hash tables, we need to store them in such a way that
they are not constrained by the VM page size.

Have (*vm.prog).mkhash store hash table values directly
in slab memory and stick raw vmrefs into the hash table
rather than encoding values relative to the scratch buffer.
We always have to re-compile the SSA code (and therefore
re-compute the hash table) when the symbol table changes,
so it is safe to use the symbol table slab memory for this
purpose.