ion: Symtab: use toindex LUT only as a cache
Since we tend to re-use symbol tables and Reset() them
periodically, the Reset() call can be a hot path due to
clearing the ion.Symtab.toindex hash table.

We can avoid clearing the toindex hash table if we let
old entries continue living in it and just double-check
after performing a look-up that the corresponding entry
in the symbol table is actually the one that is supposed to be there.