ion: Chunker: use a higher compress threshold for interning
Move the repetition threshold from 3 to 10 repetitions.
This changes the number of symbols interned in each span
of dev logs from ~13,500 down to ~700 while only increasing
the decompressed size of the data by about 2.5%.

See also #2330: there are nontrivial costs to interned strings,
so we need to be pretty certain that interning them is profitable.

size before:
399507456 Jan 18 20:29 /home/phil/core.0.ion
size after:
408944640 Jan 18 22:40 /home/phil/core.10ct.ion