ion: improvements to the Field type (#2347)
This adds some ergonomic improvements to the Field type in anticipation
of using it more widely across the codebase. In particular this adds a
ReadField function to handle reading struct fields, and embeds the Datum
field (formerly named Value) allowing a Field to be used lexically more
like a Datum.