ion: introduce Bag for sets
This change introduces ion.Bag for representing groups
of ion datums, and it changes expr.Member to use ion.Bag
for the set of elements to match.

Additionally, the old chacha8x4 code has been rehabilitated
for use during tree construction, as hashing is a relatively
expensive part of constructing the semi-join hash-set.

This makes large semi-joins significantly faster, and it
also fixes a bug in which certain semi-joins were not matched
correctly due to the presence/absence of symbols in struct fields.