ion: change Datum accessors to produce type errors (#2349)
This changes methods for accessing the value underlying ion datums to
return a TypeError instead of a boolean value, as well as adds methods
for checking the type of a datum. To prevent unnecessary allocations,
callsites have been updated to check the types before calling the
accessor methods, rather than calling them and checking if err != nil.