vm: sort: refactor ktop to allow for auxvals
Allow auxvals to show up in ktop ORDER BY results
by changing the way values are computed and snapshotted
as part of ktop construction.

The ion.Datum value keeps track of originating symbol tables
for us, so we can drop the "symbol table ID" hack and just
pass around these datum values directly. The only wrinkle
is that Datum construction is somewhat expensive, so we
have to defer construction until we know that we'd actually
like to insert a value into the heap. So, the heap now returns
the *destination* for the captured datum if and only if the
value should be captured.