plan/pir: add support for optimizing grouping by partitions
This change introduces the concept of partitioning to pir.UnionMap
and performs one relatively simple optimization on GROUP BY
where we elide grouping operations if the rows have already been
grouped by a partition. (We could also do something similar for DISTINCT.)

The semantics of "UNION MAP table PARTITION BY ..." is similar to UNION MAP
except that instead of the table being partitioned along any arbitrary axis,
the table is split along a set of distinct column values known up front.

There's no execution-time support for recognizing partitioned UnionMap steps
so the optimization is only turned on for testing right now.

This is a prerequisite to performing shuffle joins on partitions.