vm: ktop: implement OFFSET
Implement OFFSET for ktop sorting (see #2357).

The implementation is fairly naive in that it
computes the entire LIMIT+OFFSET-sized heap
and only eliminates entries during finalization.
A slightly more clever solution would maintain
a min-heap and evict offsets up to OFFSET as data
was streamed in, but that's a considerably larger change,
and right now we don't allow LIMIT+OFFSET to exceed 10000 anyhow.