vm: precompute hash tables
It is often the case that the keys used in hash tables
are "immediate" ion values (strings, integers, etc.),
and in those cases we can avoid re-computing the hash table
during recompilation because its layout does not depend on
the symbol table. (For hash lookups we still have to re-encode
the *values* part of the hash table, but the key hashes remain the same.)

Also, use recompile() for the unnesting code to ensure that
it can take full advantage of precomputed tables.