blockfmt: Converter: accelerate prepending
For prepending complete blocks, we already use
server-side copy wherever possible.
However, for *incomplete* blocks, we currently re-read
all of the ion data only to compress most of it again.

Instead of passing the existing ion data through ion.Chunker,
copy the *compressed* data verbatim to the output stream as
long as each 1MB chunk of data is "full" (doesn't end in a large nop pad).
Once we encounter an under-full chunk, we resume the "slow" transcoding
of the existing data before appending the new data.
This lets us avoid both the transcoding overhead (~800MB/s) and the
compression overhead (~80MB/s) of re-ingesting the data.
Since blocks are (up to) 100MB in size, that's more than 1 second
of CPU time spent prepending that we can elide.