aws/s3: fix Prefix.VisitDir on "." (#2405)
(*Prefix).VisitDir was not handling "." correctly and was treat it as
part of the prefix when listing.

This also changes how fsutil.WalkGlob handles opening files: the Opener
interface has been removed and instead any fs.DirEntry that also
implements fs.File will be passed to the walk callback directly. This
saves on an unnecessary listing call to check prefixes for existence
when they are already known to exist.