vm: Added Type|L and hLen to a Value register (#2475)
Increases the size of RegV by 32 bytes so TLV byte and header size
can be stored in the register. The reason for storing these in RegV
is that most of the time when a value is decoded from ION data it's
required to decode these fields as well, so we can provide them at
virtually no cost.

The benefit of this change is that when TLV and header length is
known it's much easier to unbox such value, or to check its type.