plan/pir: support multiple PARTITION BY columns
Add support for multiple PARTITION BY columns
by turning multi-column hash lookup conditions
into a single-column hash lookup with a constructed
list of all the elements that need to be looked up.

We need multi-column PARTITION BY support in order
to support COUNT(DISTINCT ...) with a multi-key GROUP BY,
since we convert the latter to the former.