vm benchmarks: don't versify if input is small
When the input is small, then time to versify it
might be really big. For instance, when input contained
only short records like `{"x": 5}` (it's 4 bytes of
Ion), then the number of calls to versify were 16M and
the total time to startup a benchmark was ridiculous,
even on a fast machine.