vm: skip filtering if condition got simplified
A condition in WHERE clause might be simplified in runtime
to either true or false. It happens when the condition
refers to non-existing/existing field. For instance 'field=1'
would simplify to false if 'field' is not present in a symtab.
Likewise, if nonexisting field is queried with `is missing`
then the condition simplifies to true.