plan/pir: construct build-side dereferenced fields as lists
If we produce the build side of a join as structures, then
we can never apply the hash-const optimization in the vm,
as the values will always depend on the symbol table.

If we construct lists instead of structures to hold the
dereferenced fields on the build side of the join, then
we could enable the hash-const optimization for the *values*
of the join. (It's also slightly faster to index fields in a list
versus fields within a structure.)

This change teaches the query planner how to construct the lists,
but it doesn't include the hash-const optimization in the vm
for lookup table values.