vm: try to make hash tables constant everywhere
When hash table keys *and* values are constant with
respect to the symbol table, create a shared hash table
that can be used by all the threads.

Since the hash table values have to live in vm memory,
we need to record that allocation inside vm.prog and
have the caller free it via prog.reset.
Typically this is done in vm.QuerySink.Close.

As part of this change, the projection code was changed
to use the same early-compilation logic as the rest of
the physical operators so that it can take advantage
of the same optimizations w.r.t. recompilation, etc.