ion/blockfmt: allow concatenation of small objects
It's possible that one item in a list of objects within
the same partition is "small" (i.e. under 50MB), and sometimes
it is small enough to be less than S3's minimum part size (5MB),
which means we can't upload it as part of a regular multi-part copy.

Since we only expect 1 object within each concat group to be
this size, and since we expect it to be the "latest" object,
we can just buffer it in memory and prepend it to the trailer
before it is written out.

Fixes #2520