db: gc: improve performance
Improve GC performance by avoiding recursive directory listing.

In the "inputs-*" case, we know these files live only in one
directory, so the optimization is trivial.

In the "packed-*" case, we can use the list of full paths
in the index to determine all the directories that need to
be searched in order to track down garbage. Just collect these
up front and do a naive listing within those directories.
(We may leak garbage in partitions that never ingest anything
successfully, but that seems like an okay trade-off.)