db: implement incremental gc
Store a cursor in blockfmt.Index.UserData for scanning packfiles.
Packfile scanning is time-boxed (configurable) so that it automatically
stops after a certain amount of time (but it always scans at least one partition).

Also, move db.IndexCache into db.tableState and un-export it.
This simplifies some of the function signatures in sync.go