db: limit background scan time; respect skip_backfill
Fixes two issues (see #2555):

1. We should respect skip_backfill when deciding whether
or not to start a background scan.

2. We should limit the amount of time that scan() can spend
doing non-productive listing. If we are backfilling data from
a bucket for which we have already ingested most of the original
data, then scanning will spend most of its time deduping the input
files and ignoring them. We should instead bail out after a certain
number of seconds and just update the scan cursors. Otherwise we'll
scan forever trying to reach the batch insert threshold.