cmd/sdb: add query applet
Add 'sdb query' that works similarly to "sneller ..."

Packfiles can be queried as tables using read_file();
we may add other builtins down the road for other data sources.

The read_file() builtin respects -root, even if -root is an S3 bucket,
so the behavior is exactly the same regardless of the media on which
the data is stored. (When -root is specified, tables can also be queried
as usual, and the results are cached in /tmp or similar.)