ion: chunker: restore symbol table after failed compress
In the extremely unlikely event that the current chunk
in ion.Chunker.compress actually grows rather than shrinks
due to the symbol table being serialized, then make sure to
also restore the symbol table to its previous state.
Otherwise, the call to ion.Chunker.Flush may panic when it
sees that there are un-comitted symbols in the symbol table.
I *think* this is the root cause of #2641, but without a repro
I am left guessing.

Updates #2641