db: unify queue and sync logging
Rework the logging in db/sync.go to be a little less chatty
and enable everything by default. There's also some minor
refactoring to ensure that log lines are emitted with an
associated db/table where appropriate.

Additionally, make sure we identify each and every rejected object
so that we can post-process them as necessary.