vm: elide recompilation for constant LUTs
Don't set vm.prog.literals unconditionally when we encounter
a hash table or hash set in (*prog).symbolize.
If the hash table is fully precomputed, then it doesn't depend
on the input symbol table and we can avoid setting prog.literals,
which in turn means we can avoid some recompilation overhead.