vm: projection: marshal symbol table lazily
Currently, projector.symbolize() unconditionally re-encodes
the current symbol table so that subsequent calls to projector.writeRows()
can append data to the output buffer. However, if projection occurs after
a filtering operation that removes many rows (as it often does), then there
can very often be many more calls to symbolize() than writeRows(), which means
we waste a bunch of time encoding symbol tables that are never actually used.

Instead, defer symbol table encoding (alignedWriter.setpre) until we know
that more than zero rows of data are actually going to be projected with
the current symbol table.

This saves about 3% end-to-end for the query in #2682.