vm: use best-fit allocation for bytecode stack
The existing stack allocator does not re-use regions
of memory for allocations that are not the same size,
which means that e.g. a hash slot allocation region
cannot be re-used for two scalar slot allocations later
when the hash slot is dead.

Change the stack allocation back-end to use a best-fit allocator
that tracks every span of used/free memory on the stack.
The allocator is entirely agonistic to the size of the allocations
it is asked to perform, so it just tries to pack every new allocation
inside a "hole" in the existing used regions before extending the stack.

Also, fix two bugs revealed by this change:

 - bchashvalueplus passed the wrong arguments to hashimpl_tail()
 - geotile{x,y} were incorrectly labeled as having stBool return values

On average this reduces stack usage in the TestQueries harness
by about 30%.