vm: add support for consuming zion directly in most cases
Rather than having the zion package re-construct ion structures
completely and then re-computing their positions with the scan()
routine, we can instead just project zion fields directly into
auxiliary bindings. (For now, we only do this when the number of
fields to be projected is between 1 and 7 fields, inclusive.)

Sticking values directly into auxvals elides the scan() routine,
a data copy in zion.zipfast, and also eliminates gather overhead
in findsym_tail (since bcauxval doesn't use gathers).

In practice this improves end-to-end throughput by 0% to 20%
depending on how many fields need to be fetched, how large they are,
and so forth. Typical improvements are about 10% end-to-end.