vm: zion.go: use a fixed stride for columns
Rather than traversing the shape to find the row count up front,
used a fixed column stride and write out rows incrementally as they
are processed.

In practice this makes things 5-10% faster, as zll.Shape.Count()
is moderately expensive, and also the assembly is slightly neater
when we use a fixed stride width with a single output slice.