vm: extend bytecode generator with opcode analysis
Analysis relies of use in the code predefined macros.
It checks:

- if slots are accessed correctly (proper offset and size);
- if GPR registers holds a slot when loading/storing data;
- if next advance offset is calculated correctly.

In the case of opcodes that have variable number of arguments
analysis is not possible, as some offset calculations are done
outside the macros.

The tool found two actual mistakes in code, and a few
inconsistencies in using `IN`/`OUT` markers; fixed in this commit.