iguana: reduce heap allocations in compression
Reduce the number of heap allocations performed in the
compression path. Stick all the temporary state for encoding
inside iguana.Encoder so that things like the hash chain table,
temporary ANS buffers, etc. can be re-used.
The ANS code also gets the same treatment.

The largest remaining source of allocations in the encoder path
is the hash chain table, since *most* of the hash chain slices
end up being nil. It might be worth investigating making this
a real "chain" at each table position in order.

This change roughly triples the performance of BenchmarkCompress
on the one piece of reference data.