iguana: add test for oob writes; shift encoding rules
Add a check in unit tests for out-of-bounds writes
past the end of the destination buffer and fix the
assembly code so that the second half of the zmm register
used for copying doesn't extend past the end of the input.
Using 32-byte instead of 64-byte copies does not seem
to negatively impact performance on any of our reference
datasets.

Also, change the encoding algorithm slightly so that we
can allow matches to extend up to the very end of the source
buffer, which means there can be a zero length literal at
the end of the encoding stream. Fix the portable decoder
so that it handles zero-length literals correctly.

The new (more permissive) match checking claws back
all of the compression size loss from eliding duplicate
matches for matches longer than the simd block size.
In many cases the compression ratio is actually *better*
than before this patch set.