vm: unify handling split for aggregates with state
Aggregates that use more bytes of state than two scalar values
get the common path of execution.

A bytecode program that runs on a child machine is executed normally,
only when saving its final value, we store its entire aggregate buffer
as a string.

A program on the main node is slightly modified. All aggregates
that need a bigger internal state are handled by a single opcode
that only stores the string location the within input data; we
process the incoming records row-by-row. Then execution gets back
to the Go code and we're performing merge operation on the
buffers for these aggregates.