aws/s3: ignore context for readDirAt results
This is a hacky work-around fix for a bug to buy a
little time to come up with a better solution.

The context we use for listing in db/scan.go is
deliberately short-lived (15 seconds) in order
to cause listing to bail out quickly if there
are lots of dirents to traverse. However, that
deadline also ends up being part of the returned
fs.File, which means that if the batch itself takes
longer than 15 seconds to ingest, then the whole
batch fails.

For now, just ignore the parent context when
yielding an fs.File in s3.BucketFS.VisitDir.