vm: projection: adjust bytecode.auxpos correctly
The vm takes care of adjusting bytecode.auxpos for each
input lane, but projection sometimes has to bail out
partway through processing the 16 lanes if there isn't
enough space in the output buffer. In that case, we need
to make sure that we only advance the number of lanes
that were actually projected. The easiest fix is simply
to save auxpos before invoking the bytecode and then
restoring it to the previous value plus only the number
of projected lanes.