iguana: ensure streams are padded
Since we load 32 bytes from the literals stream
unconditionally as part of match+literal copy-out,
we need to ensure that it is padded appropriately
for a load of that width.

As a matter of precaution, ensure that the other streams
are padded out to their corresponding load widths in
the assembly decompressor as well.