plan, plan/pir: do not emit aggregates from empty parts
If we turn an aggregation into a partitioned aggregation,
we need to deliberately avoid emitting rows in the final aggregation
that correspond to partitions that had zero rows reach the aggregation
in the first place.

Add a "NonEmpty" tag on aggregate operations that causes them to
avoid flushing any output when no rows appear on the input, and
teach the query planner to use NonEmpty aggregates as necessary.

Also, add a extra-parts tag to the testquery harness so that
we can get some coverage of empty partitions in vm.TestQueries.

Fixes #2978