tenant: manager: relax handling of tnproto.Key values
Rather than delivering a hard error when a (ID, Key) pair
isn't the one we expect, just create a new tenant process
with the associated tuple. This still gives us the same privacy
guarantees as delivering an error, but it also allows us to gracefully
handle a tenant being torn down and then set up again in rapid succession.

Fixes #3033