iguana: add support for any match length
Allow any offset+matchlen pair to be handled safely
by the assembly decoder as long as the final write
for copying out the match does not extend past the
end of the target buffer.

This improves compression ratio substantially
for objects that are highly compressible,
and has negligible impact on the performance
of decompression in general.