plan: Leaf: re-order filtering and partitioning
plan.Input.Partition can exploit the partition cache
only if the plan.Input has a cache in the first place.
plan.Input.Filter produces a copy of the input (without a cache)
so the partition cache couldn't be taken advantage of when a
predicate was present.

Re-order plan.Input.Filter and plan.Input.Partition so that
partitioning happens first.