vm: Fixed BOOL_AND() to not treat non-bool as FALSE (#3053)
The issue was that the common path of BOOL_AND() was coercing to BOOL
by istrue.v before calling the aggregator. This caused non-BOOL values to be
coerced to FALSE, which caused BOOL_AND() to result in FALSE.

This fixes this by avoiding the coercion at the higher level, leaving
the decision on SSA code - basically if the input is a value, it would
unbox bool to i64 instead, and AND aggregate i64 instead, which yields
the correct output (BOOL_AND and int64 AND aggregations are essentially
the same, just boxed differently to present the final result).