vm: fix findbc scratch handling
The previous "fix" for scratch handling in findbc was incorrect.
The scratch buffer should only be reset once per invocation.
The reason we were running out of scratch space is that the
scratch usage calculation assumes we are only going to process
sixteen lanes at a time, but bcfind may use scratch space for
more than sixteen lanes (in our case, len(delims)*9 bytes, so up to 256*9 bytes).

Ensure that compilation always uses the max page size for ORDER BY findbc scratch space
and add a unit-test that blows up when this is handled incorrectly.