zion: add "iguana/specialized" compression and hints
Add scaffolding for ion-specialized compression.

Add a new compression "type" that is just default-iguana
plus specialized encoding on a per-bucket basis as determined
by the first byte of the compressed bucket.

Add zll.BucketHints and have it passed to each invocation of
the compressor. For now it just tracks trivial stuff like the
number of elements and the union of all the types of the elements.