plan: Server: clear writeFailed at initialization time
If a query is canceled and causes the plan.server.writeFailed flag
to be set, we don't clear it before re-using the server, which I believe
is the root cause of #3063.

Also, improve logging of invocations of plan.Server.Serve in ProxyExec
calls so that we capture query errors on both the sending and receiving
end of a split query plan.