jsonrl: eisel_lemire: allow insiginificant imprecision
Allow the Eisel-Lemire code to produce imprecise results
when the result would be equivalent under decimal rounding.
Semantically speaking, we don't know what the insignificant bits
of float64-parsed decimal results would be anyway, and this still
preserves the invariant that the parsed text and the result text
in JSON conversion are identical.