bugfix in bcContainsSuffixUTF8 (#3168)
Bugfix in bcContainsSuffixUTF8.

There was an issue where a needle is being incorrectly matched in a situation where the needle contains four trailing null characters, while the data does not have such characters. When loading the data, if a register is too large to contain the contents of the last bytes of the data, the remaining positions are filled with byte 0, which causes these bytes to be mistakenly matched.

This has been fixed by loading the remaining lower bound of the length of data that can result in a match. This approach also enables a faster realization that even if a string matches for the first several characters, the remainder cannot match because the minimum required length to achieve a match is too large for the remaining number of bytes in the data.