Bugfix evalbc test (#3169)
bugfix: only when expected and observed length AND expected and observed offsets are unequal, problems are reported. Obviously, issues should be reported when either length or offsets are incorrect. This bug only prevented unit tests (manual) from being reported. Brute-force tests and fuzzy tests were not affected.

A lot of unit tests (UT1/UT2) had to be updated to work with vm displacements (that were now observed as false positives)