Faster UTF8 string compare (#3172)
Better usage of needle and data length to allow quick exit for UTF8 string comparisons.

For example, when we are searching for a needle like, "ssss𐍈𐍈", any data with less than 12 bytes (1 bytes for each 's' and 4 bytes for each '𐍈') cannot possibly contain our needle. These 12 bytes set the minimum requirement for the data's length. If we take the example of the 10-byte string "ſſſſſ" (note that 'ſ' is a two-byte alternative to the ASCII 's) the string will be considered for a match.

Currently, our substring code doesn't take into account the byte counts. The updated implementation considers the lower bound requirement at every position within the needle.